backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.11.0
duckdb==0.9.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
crewai==0.36.1
python-dotenv==1.0.0
pydantic>=2.7.0,<3.0.0
asyncio-mqtt==0.16.1
faker==20.1.0
python-multipart==0.0.6